Version: 0.3.21. WiP. It's coming passwords, cards and more

+++ frontend/app/lib/queries/delete2fa.ts
@@ -1,11 +1,22 @@
 "use server";
 
-import { PasswordValidator } from "@/app/lib/schemas.z";
+import { PasswordSchema } from "@/app/lib/schemas.z";
+import { z } from "zod";
+import getTranslations from "../helpers/getTranslations";
 import getUserDataJson from "../helpers/getUserDataJson";
 import protectedQuery from "../helpers/protectedQuery";
 
-export default async function delete2fa(data: PasswordValidator) {
-  // TODO: Something is wrong with the session access_token in server side
+export default async function delete2fa(data: z.infer<typeof PasswordSchema>) {
+  const t = await getTranslations();
+  const validatedData = PasswordSchema.safeParse(data);
+  if (!validatedData.success) {
+    return {
+      success: null,
+      error: validatedData.error.message,
+      userData: null,
+    };
+  }
+
   const result = await protectedQuery({
     url: `/2fa/delete_totp_device/`,
     method: "DELETE",
@@ -13,43 +24,44 @@ export default async function delete2fa(data: PasswordValidator) {
       "Content-Type": "application/json",
       Accept: "application/json",
     }),
-    body: data,
+    body: validatedData.data,
     credentials: "include",
   });
 
   if (result instanceof Error) {
     return {
       success: null,
-      error: "profile.twofactor.notDeleted",
+      error: t("error.somethingWentWrong"),
       userData: null,
     };
   }
 
-  if (result?.response.ok) {
-    // TODO: apply new access token without TOTP device to session
-    const { access, refresh } = (await result.json) as {
+  if (result?.ok) {
+    const { access, refresh } = (await result.json()) as {
       access: string;
       refresh: string;
     };
+
     const updatedUserData = await getUserDataJson(access, refresh);
+
     return {
-      success: "profile.twofactor.deleted",
+      success: t("profile.twofactor.deleted"),
       error: null,
       userData: updatedUserData,
     };
   }
 
-  if (result?.response.status === 400) {
+  if (!result?.ok) {
     return {
       success: null,
-      error: "profile.twofactor.notDeleted",
+      error: t("profile.twofactor.notDeleted"),
       userData: null,
     };
   }
 
   return {
     success: null,
-    error: "profile.twofactor.notDeleted",
+    error: null,
     userData: null,
   };
 }
